Rename Instance Controller to LinuxConsumptionInstanceAdmin Controller
