+++ Kudu.Services.Web/Startup.cs
@@ -91,9 +91,10 @@ namespace Kudu.Services.Web
                 .AddApplicationPart(kuduServicesAssembly).AddControllersAsServices()
                 .AddApiExplorer();
 
-            // Instance authentication and authorization for running in service fabric mesh
-            services.AddInstanceAdminAuthentication();
-            services.AddInstanceAdminAuthorization();
+            // Add middleware for Linux Consumption authentication and authorization
+            // when KuduLIte is running in service fabric mesh
+            services.AddLinuxConsumptionAuthentication();
+            services.AddLinuxConsumptionAuthorization();
 
             services.AddSwaggerGen(c =>
             {
@@ -455,10 +456,10 @@ namespace Kudu.Services.Web
 
                 // Instance administration
                 routes.MapRoute("admin-instance-info", "admin/instance/info",
-                    new {controller = "Instance", action = "Info"},
+                    new {controller = "LinuxConsumptionInstanceAdminController", action = "Info"},
                     new {verb = new HttpMethodRouteConstraint("GET")});
                 routes.MapRoute("admin-instance-assign", "admin/instance/assign",
-                    new {controller = "Instance", action = "AssignAsync" },
+                    new {controller = "LinuxConsumptionInstanceAdminController", action = "AssignAsync" },
                     new {verb = new HttpMethodRouteConstraint("POST")});
 
                 // Live Command Line
