+++ Kudu.Tests/Services/Infrastructure/Authentication/ArmAuthenticationHandlerTests.cs
